friend/requirements.txt
Django
djangorestframework==3.14.0
